+++ NEWS.md
@@ -5,7 +5,7 @@
 * Simplify Rcpp glue: remove unnecessary `as<>()` calls (ec4e51a).
 * Simplify trajectory's head/tail management (06432a8).
 * Ongoing (unfinished) arrivals are reported with `get_mon_arrivals(ongoing = TRUE)` (#73).
-* Now, `run(until)` runs the simulation exactly until `until`, instead of until the first event scheduled at a time >= `until`.
+* Now, `run(until)` runs the simulation exactly until `until`, instead of until the first event scheduled at a time >= `until` (e7264f6).
 
 # Changes in version 3.4.1